rayw: order of the tests / components
in outside-in TDD, the order in which you finish the components might be
even more important than in classicist TDD.

you get a false sense of completion, having distributed the
responsibilities into several components. now it is but a walking
skeleton, as the components do not properly work on their own.

the list with the pending feature might be more needed than in classicist.
maybe it is my own lack of mastery who impedes / causes this.