rayw: primitive obsession
the representation for cash, units is an int; this does not allow for
behaviour to stick to these types.

also, there is a lack of symmetry, as some type Cost has already been
created for representing this, but is not used yet.

this refactoring, i consider less important. it would be some kind of
accidental in the scale of Brooks.

@trikitrok told me about "structural" refactoring, as opposed to another
kind. i don't recall what it was about, but a proposal would be:

  * structural: affects the structure of the code
  * decorative: affects only the look and feel of the code

 #rayw #primitiveObsession